Added support for error notifications.

+++ apps/erbium/erbium.h
@@ -232,7 +232,7 @@ struct rest_implementation {
   int (* get_post_variable)(void *request, const char *name, const char **value);
 
   /** Send the payload to all subscribers of the resource at url. */
-  void (* notify_subscribers)(resource_t *resource, uint16_t counter, void *notification);
+  void (* notify_subscribers)(resource_t *resource, int32_t counter, void *notification);
 
   /** The handler for resource subscriptions. */
   restful_post_handler subscription_handler;
